add initial full impl of training script with args
this commit finalizes the initial version of the training script with
support for fixed sized input gifs only. variables window size support,
as well as variable gif dim support for FCN is yet to be added. This
version is inteded for testing of the initial archs for one to one frame
input-target frame compression

Some arguments in the cmd argparser were also tweaked to fit training
needs