add new dataset reader
this dataset reader comprehensively covers the different usecases and
methods of reading the GIF dataset based on input arguments at
construction time.

will crop and batch the inputs and return them as np.ndarray with
dtype=np.float16. palette will be returned with a dtype of np.int8

output batches have the shape (batch_size, height, width) if window size
is 1 and shape (batch_size, window_size, height, width) is window size
is > 1. currently the channel dimention is removed from the output
batches which may lead to issues with the placeholders in the trianing
file. This remains yet to be tested.

What has been tested is the basic fuctioanlity of batch generation and
frame marching, cropping and pruning of the dataset based on the
arguments. These tests are preliminary, and may not be airtight.

dataset now also contains static methods for non-parameterised GIF
dataset access. these are mainly intended for inference since we do not
need dataset curation, batch generation etc during that time.