add support for variable size GIF dims during training for FCN
FCNs will accept any input size, this commit adds support for generating
training batches from the dataset that do not crop the GIF before
returning it.

One contraint for this is that the batch_size must be set to 1. This is
because otherwise we would have to pad the output array and we do not
want to send padded inputs to the FCNs if possible. This may need to
change in the future which would allow for batch_size > 1 but this is an
initial solution for now.