added transform to dataset reader. closes GH-3
consumer of data during training of the network may want to apply a
transform to the data right before feeding it to the network. This
commit adds an option to give the dataset reader a callable that the
user can supply if they wish. This function will simply be applied to
'frame_batch' and 'target_batch' before being yield-ed by the datset generator.

The motivation for this to be a functor passed to dataset instead of
train.py doing this itself if two-fold. i) composability, and ii) single
responsibility. train.py should not have to handle data manipulation and
dataset reader should not worry about all the possible transforms.