add transform configuration support and basic transforms
this commit adds the ability to apply normalize and denormalize input
data to the encoder and output predictions from the decoder. train.py
accepts an input transform and passes it to the dataset reader as a
functor if it is specified.

argument parser is changed to include the cmd args that can be used to
specify the name of the input and output transforms to be used. seperate
arguements are configured for training and inference.

addresses #2