Restructured train graph. Addresses GH-13 and GH-12
The training graph is "split" in two parts by having a placeholder for Z (Z_placeholder) in addition to Z output from encoder. This is done to remove the dependency of T_hat graph node on Z (output from encoder) node, and ultimately X (input frames) node. Session runs are modified to accomodate for this change. For inference, encoder and decoder subgraphs are extracted, variables are converted to constants, and pb's are saved.

compress.py and decompress.py are modified to load (sub)graph definitions from pb's. Input and output nodes are extracted from loaded graph to fetch sess.run() i.e. "do the inference"